python/requirements.txt
fastapi
uvicorn
pillow
tesserocr
openai
python-multipart